requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
apscheduler==3.10.4
